# openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "灵魂旅游",
    "description": "",
    "version": "1.0.0"
  },
...
  "paths": {
    "/moments/{momentId}": {
      "post": {
        "summary": "获取单条动态详情",
        "deprecated": false,
        "description": "",
        "tags": [],
        "parameters": [
          {
            "name": "momentId",
            "in": "path",
            "description": "",
            "required": true,
            "example": "",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-CH",
            "in": "header",
            "description": "",
            "required": false,
            "example": "",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "moment_id": {
                    "type": "string",
                    "format": "uuid"
                  }
                },
                "required": [
                  "moment_id"
                ]
              },
              "examples": {}
            }
          }
        },
        "responses": {
          "200": {
            "description": "未找到",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MomentDetail"
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
...
    "/moments/{momentId}/comments": {
      "post": {
        "summary": "发布评论",
        "deprecated": false,
        "description": "",
        "tags": [],
        "parameters": [
          {
            "name": "momentId",
            "in": "path",
            "description": "",
            "required": true,
            "example": "",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-CH",
            "in": "header",
            "description": "",
            "required": false,
            "example": "",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "user_id": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "moment_id": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "content": {
                    "type": "string"
                  }
                },
                "required": [
                  "user_id",
                  "moment_id",
                  "content"
                ]
              },
              "examples": {}
            }
          }
        },
        "responses": {
          "200": {
            "description": "参数错误",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 400
                    },
                    "message": {
                      "type": "string",
                      "example": "参数错误"
                    },
                    "data": {
                      "type": "null",
                      "example": null
                    }
                  }
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "MomentDetail": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "user": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string"
              },
              "nickname": {
                "type": "string"
              },
              "avatar_url": {
                "type": "string"
              }
            },
            "required": [
              "id",
              "nickname",
              "avatar_url"
            ]
          },
          "content": {
            "type": "string"
          },
          "image_urls": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "location": {
            "type": "string"
          },
          "created_at": {
            "type": "string"
          },
          "likes_count": {
            "type": "integer"
          },
          "comments_count": {
            "type": "integer"
          }
        },
        "required": [
          "id",
          "user",
          "content",
          "image_urls",
          "location",
          "created_at",
          "likes_count",
          "comments_count"
        ]
      },
...
    },
    "securitySchemes": {}
  },
  "servers": [],
  "security": []
}